requirements.txt
playwright
python-dotenv
pyotp